feat: Add functionality to upload and store data sources
The code changes include adding a new function `add_data_sources` that allows users to upload JSON files as data sources. The uploaded files are stored in the `data` directory. This functionality is added to the `setup.impl.jac` file.

This commit message follows the established convention of using a verb in the imperative form to describe the purpose of the code changes.